refactor(billing,gemini): 抽取 pricing/compat 剩余 TK 内联块
billing registry/image 与 gemini prepare/oauth-cooldown 迁入 companion。
功能不变。no-web-impact
